+++ gatsby-config.js
@@ -1,6 +1,8 @@
 // Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
+require('dotenv').config();
+
 const path = require('path');
 
 module.exports = {