+++ gatsby-node.js
@@ -4,3 +4,4 @@
 exports.createPages = require('./util/gatsby/createPages');
 exports.onCreateNode = require('./util/gatsby/onCreateNode');
 exports.onCreatePage = require('./util/gatsby/onCreatePage');
+exports.onPostBuild = require('./util/gatsby/onPostBuild');