+++ src/locales/en/search.js
@@ -1,0 +1,9 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+export default {
+  searchPagesTitle: 'Pages',
+  searchPlaceholder: 'Search',
+  searchResult: 'result',
+  searchResults: 'results',
+};