+++ src/locales/fi.js
@@ -10,6 +10,7 @@ import indexMessages from './fi/index';
 import metaMessages from './fi/meta';
 import navigationMessages from './fi/navigation';
 import notFoundMessages from './fi/not-found';
+import searchMessages from './fi/search';
 
 export const messages = {
   ...blogMessages,
@@ -19,6 +20,7 @@ export const messages = {
   ...metaMessages,
   ...navigationMessages,
   ...notFoundMessages,
+  ...searchMessages,
 };
 
 export const lang = defineMessages(messages);