+++ src/locales/fi/search.js
@@ -1,0 +1,9 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+export default {
+  searchPagesTitle: 'Sivut',
+  searchPlaceholder: 'Haku',
+  searchResult: 'tulos',
+  searchResults: 'tulosta',
+};