+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.8.0] - 2021-03-12
+
+### Added
+
+- Search box to the site.
+- Automatic indexing of the pages of the site on build.
+
 ## [0.7.2] - 2021-03-11
 
 ### Changed
@@ -180,7 +187,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.7.2...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/visiosto/visiosto.fi/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/visiosto/visiosto.fi/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/visiosto/visiosto.fi/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/visiosto/visiosto.fi/compare/v0.6.2...v0.7.0